Newtonsoft vs System.Text.Json. Testing models using Newtonsoft
