--- src/backend/integration/Fusion.Resources.Integration.Models/Events/ResourceAllocationRequestEvent.cs
@@ -1,17 +1,0 @@
-﻿using System;
-
-namespace Fusion.Resources.Integration.Models.Events
-{
-    public class ResourceAllocationRequestEvent
-    {
-        public ResourceAllocationRequestEvent(Guid requestId, Guid positionId, Guid instanceId)
-        {
-            this.RequestId = requestId;
-            this.PositionId = positionId;
-            this.InstanceId = instanceId;
-        }
-        public Guid RequestId { get; set; }
-        public Guid PositionId { get; set; }
-        public Guid InstanceId { get; set; }
-    }
-}
