--- src/backend/integration/Fusion.Resources.Integration.Models/Events/ResourceAllocationRequestEventType.cs
@@ -1,8 +1,0 @@
-﻿namespace Fusion.Resources.Integration.Models.Events
-{
-    public enum ResourceAllocationRequestEventType
-    {
-        RequestCreated,
-        RequestRemoved
-    }
-}
