--- src/backend/integration/Fusion.Resources.Integration.Models/Events/ResourceAllocationRequestSubscriptionEvent.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Text.Json.Serialization;
-
-namespace Fusion.Resources.Integration.Models.Events
-{
-    public class ResourceAllocationRequestSubscriptionEvent
-    {
-        public Guid ItemId { get; set; }
-
-        public ResourceAllocationRequestEvent Request { get; set; }
-
-        [JsonConverter(typeof(JsonStringEnumConverter))]
-        public ResourceAllocationRequestEventType Type { get; set; }
-    }
-}
